fix exporter test

+++ optimum/exporters/onnx/config.py
@@ -302,7 +302,9 @@ class EncoderDecoderOnnxConfig(OnnxSeq2SeqConfigWithPast):
             encoder_onnx_config_constructor = TasksManager.get_exporter_config_constructor(
                 exporter="onnx", task="feature-extraction", model_type=config.encoder.model_type
             )
-            self._encoder_onnx_config = encoder_onnx_config_constructor(config.encoder, preprocessors=preprocessors)
+            self._encoder_onnx_config = encoder_onnx_config_constructor(
+                config.encoder, int_dtype=int_dtype, float_dtype=float_dtype, preprocessors=preprocessors
+            )
             self._normalized_config.ENCODER_NORMALIZED_CONFIG_CLASS = self._encoder_onnx_config._normalized_config
 
         if self._behavior is not ConfigBehavior.ENCODER:
@@ -323,7 +325,7 @@ class EncoderDecoderOnnxConfig(OnnxSeq2SeqConfigWithPast):
                 )
 
             self._decoder_onnx_config = decoder_onnx_config_constructor(
-                config.decoder, preprocessors=preprocessors, **kwargs
+                config.decoder, int_dtype=int_dtype, float_dtype=float_dtype, preprocessors=preprocessors, **kwargs
             )
             if issubclass(decoder_onnx_config_constructor.func, OnnxSeq2SeqConfigWithPast):
                 self._decoder_onnx_config = self._decoder_onnx_config.with_behavior(
